refactor: rename utils to route-utils
